+++ mobile/android/android-components/components/support/ktx/src/test/java/mozilla/components/support/ktx/android/view/ActivityTest.kt
@@ -11,17 +11,19 @@ import android.view.Window
 import android.view.WindowManager
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import mozilla.components.support.test.any
+import mozilla.components.support.test.argumentCaptor
 import mozilla.components.support.test.mock
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.mockito.Mockito.`when`
+import org.mockito.Mockito.anyInt
 import org.mockito.Mockito.mock
 import org.mockito.Mockito.never
 import org.mockito.Mockito.times
 import org.mockito.Mockito.verify
 
-@Suppress("Deprecation")
+@Suppress("DEPRECATION")
 @RunWith(AndroidJUnit4::class)
 class ActivityTest {
 
@@ -43,14 +45,58 @@ class ActivityTest {
     }
 
     @Test
-    fun `check enterImmersibleMode sets the correct flags`() {
+    fun `check enterToImmersiveMode sets the correct flags`() {
 
         activity.enterToImmersiveMode()
+
+        // verify entering immersive mode
+        verify(window).addFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON)
+        verify(decorView).systemUiVisibility = View.SYSTEM_UI_FLAG_FULLSCREEN
+        verify(decorView).systemUiVisibility = View.SYSTEM_UI_FLAG_HIDE_NAVIGATION
+        // verify that the immersive mode restoration is set as expected
+        verify(window.decorView.viewTreeObserver, never()).addOnWindowFocusChangeListener(any())
+        verify(window.decorView).setOnSystemUiVisibilityChangeListener(any())
+    }
+
+    @Test
+    fun `check setAsImmersive sets the correct flags`() {
+        activity.setAsImmersive()
 
         verify(window).addFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON)
         verify(decorView).systemUiVisibility = View.SYSTEM_UI_FLAG_FULLSCREEN
         verify(decorView).systemUiVisibility = View.SYSTEM_UI_FLAG_HIDE_NAVIGATION
         verify(window.decorView.viewTreeObserver, never()).addOnWindowFocusChangeListener(any())
+    }
+
+    @Test
+    fun `check enableImmersiveModeRestore sets the correct insets listeners`() {
+        activity.enableImmersiveModeRestore(null)
+        verify(window.decorView.viewTreeObserver, never()).addOnWindowFocusChangeListener(any())
+        verify(window.decorView).setOnSystemUiVisibilityChangeListener(any())
+
+        val windowFocusChangeListener: ViewTreeObserver.OnWindowFocusChangeListener = mock()
+        activity.enableImmersiveModeRestore(windowFocusChangeListener)
+        verify(window.decorView.viewTreeObserver).addOnWindowFocusChangeListener(windowFocusChangeListener)
+        verify(window.decorView, times(2)).setOnSystemUiVisibilityChangeListener(any())
+    }
+
+    @Test
+    fun `check enableImmersiveModeRestore set insets listeners have the correct behavior`() {
+        val insetListenerCaptor = argumentCaptor<View.OnSystemUiVisibilityChangeListener>()
+        activity.enableImmersiveModeRestore(null)
+        verify(window.decorView).setOnSystemUiVisibilityChangeListener(insetListenerCaptor.capture())
+
+        insetListenerCaptor.value.onSystemUiVisibilityChange(View.SYSTEM_UI_FLAG_FULLSCREEN)
+        // If the activity requested to enter fullscreen no immersive mode restoration is needed.
+        verify(window, never()).addFlags(anyInt())
+        verify(decorView, never()).systemUiVisibility = anyInt()
+        verify(decorView, never()).systemUiVisibility = anyInt()
+
+        insetListenerCaptor.value.onSystemUiVisibilityChange(View.SYSTEM_UI_FLAG_LIGHT_NAVIGATION_BAR)
+        // Cannot test if "setAsImmersive()" was called it being an extension function but we can check the effect of that call.
+        verify(window).addFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON)
+        verify(decorView).systemUiVisibility = View.SYSTEM_UI_FLAG_FULLSCREEN
+        verify(decorView).systemUiVisibility = View.SYSTEM_UI_FLAG_HIDE_NAVIGATION
     }
 
     @Test
@@ -64,6 +110,7 @@ class ActivityTest {
         verify(window, never()).clearFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON)
         verify(decorView, never()).systemUiVisibility = View.SYSTEM_UI_FLAG_FULLSCREEN.inv()
         verify(decorView, never()).systemUiVisibility = View.SYSTEM_UI_FLAG_HIDE_NAVIGATION.inv()
+        verify(decorView, never()).setOnSystemUiVisibilityChangeListener(null)
 
         attributes.flags = WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON
 
@@ -72,15 +119,7 @@ class ActivityTest {
         verify(window).clearFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON)
         verify(decorView, times(2)).systemUiVisibility = View.SYSTEM_UI_FLAG_VISIBLE
         verify(window.decorView.viewTreeObserver, never()).removeOnWindowFocusChangeListener(any())
-    }
-
-    @Test
-    fun `check enterImmersibleMode with callback adds the correct callback`() {
-        val callback: ViewTreeObserver.OnWindowFocusChangeListener = mock()
-
-        activity.enterToImmersiveMode(callback)
-
-        verify(window.decorView.viewTreeObserver).addOnWindowFocusChangeListener(callback)
+        verify(decorView).setOnSystemUiVisibilityChangeListener(null)
     }
 
     @Test